wip on rasterio tests
